+++ src/config_store.h
@@ -30,7 +30,6 @@ struct Settings {
   uint32_t heartbeat_ms;
   bool heartbeat_enabled;
   uint32_t ack_timeout_ms;
-  uint32_t mqtt_remote_retry_timeout_ms;
   bool tx_mqtt_remote_polling_enabled;
   uint32_t tx_mqtt_remote_default_poll_interval_ms;
   bool input_control_paired_lora_enabled;